% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtmospherePressure.R
\name{AP.kPa}
\alias{AP.kPa}
\title{Atmosphere pressure}
\usage{
AP.kPa(Tair, Rh, Elev)
}
\arguments{
\item{Tair}{Temperature in degree Celcius}

\item{Rh}{Relative humidity \link{ratio}}

\item{Elev}{Elevation in meter}
}
\value{
Atmp.pressure in kPa
}
\description{
Calculating atmosphere pressure in kPa
}
