% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declination.R
\name{declination}
\alias{declination}
\title{Declination}
\usage{
declination(Jday)
}
\arguments{
\item{Jday}{Julian date or day of the year \link{day} (i.e., 1 to 365 or 366)}
}
\value{
Solar declination \link{rad}
}
\description{
Estimating solar declination in radian
}
