% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extraterrestrial_Solar.R
\name{Extraterrestrial_Solar}
\alias{Extraterrestrial_Solar}
\title{Extraterrestrial solar radiation}
\usage{
Extraterrestrial_Solar(JD, lat)
}
\arguments{
\item{JD}{Julian day}

\item{lat}{Latitude in degrees}
}
\value{
Extraterrestrial solar radiation \link{MJ/m2/day}
}
\description{
Calculating potential solar radiation in \link{MJ/m2/day} given day of the year and latitude
}
