% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hargreaves.R
\name{Hargreaves_Day}
\alias{Hargreaves_Day}
\title{Hargreaves' PET estimation}
\usage{
Hargreaves_Day(tmin, tmax, JD, lat)
}
\arguments{
\item{tmin}{Daily minimum temperature}

\item{tmax}{Daily maximum temperature}

\item{JD}{Julian day}

\item{lat}{Latitude in degree decimal (ex: 56.1)}
}
\value{
Daily PET in mm/day
}
\description{
Estimating daily PET based on min/max temperature
}
\examples{
Hargreaves_Day(15,25,as.Date("2011-11-30"),54.0)

}
