% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Julian_Day.R
\name{Julian_Day}
\alias{Julian_Day}
\title{Julian Day}
\usage{
Julian_Day(Date.Seq)
}
\arguments{
\item{Date.Seq}{Date (format="\%Y-\%m-\%d") in "Date"}
}
\value{
Julian day
}
\description{
Converting date(YYYY-MM-DD) to Julian day
}
\examples{
BD<-as.Date('2000-01-01')
ED<-as.Date('2000-12-31')
Date.Seq<-as.Date((seq(BD,ED,by="1 day")))
Julian_Day(Date.Seq)
}
