% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Hargreaves.R
\name{Mod_Hargreave_Day}
\alias{Mod_Hargreave_Day}
\title{Modified Hargreaves' PET estimation}
\usage{
Mod_Hargreave_Day(tmin, tmax, times, lat, pr, monthlyDate, ...)
}
\arguments{
\item{tmin}{Daily minimum temperature}

\item{tmax}{Daily minimum temperature}

\item{times}{Date (ex: 1950-01-01) in Date format}

\item{lat}{Latitude in degree decimal (ex: 56.1)}

\item{pr}{Monthly precipitation \link{mm/month}}

\item{monthlyDate}{Year and month in a date format (YYYY-MM)}

\item{...}{}
}
\value{
Daily PET in mm/day
}
\description{
Estimating daily PET with the modified Hargreaves' method with min/max temperature and monthly precipitation
}
\details{
More details on the the modified Hargreaves' method, please refer to Droogers P., Allen, R., 2002,
Estimating reference evapotranspiration under inaccurate data conditions, Irrgation and Drainage System, 16, 33-45.
}
