% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Net_Out_Longwave.R
\name{Net_Out_Longwave}
\alias{Net_Out_Longwave}
\title{Net outgoing longwave radiation}
\usage{
Net_Out_Longwave(JD, Ele, lat, Tmin, Tmax, e, Rs, Rs_unit = "MJm2day")
}
\arguments{
\item{JD}{Julian day}

\item{Ele}{Elevation in meter}

\item{lat}{Latitude in decimal degrees}

\item{Tmin}{Daily minimum temperature in degree Celsius}

\item{Tmax}{Daily maximum temperature in degree Celsius}

\item{e}{Partial water vapor pressure in KPa}

\item{Rs}{Incoming solar (shortwave) radiation in MJ/m2/day or W/m2}

\item{Rs_unit}{Unit of Rs in character: 'Wm2' or 'MJm2day'}
}
\value{
Rnl \link{MJ/m2/day}
}
\description{
Net outgoing longwave radiation
}
