% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ordin.R
\name{Ordin}
\alias{Ordin}
\title{Ordin' PET estimation}
\usage{
Ordin(tmin, tmax, JD, lat)
}
\arguments{
\item{tmin}{Daily minimum temperature}

\item{tmax}{Daily maximum temperature}

\item{JD}{Julian day}

\item{lat}{Latitude in degree decimal (ex: 56.1)}
}
\value{
Daily PET in mm/day
}
\description{
Estimating daily PET based on min/max temperature
}
\details{
More details on the the Ordin's method, please refer to Oudin, L., Hervieu, F., Michel, C., Perrin, C., Andr?assian, V.,
Anctil, F. & Loumagne, C. (2005) Which potential evapotranspiration input for a lumped rainfall-runoff model? Part 2:
Towards a simple and efficient potential evapotranspiration model for rainfall-runoff modelling.
J. Hydrol. 303(1-4), 290-306
}
\examples{
Ordin(15,25,as.Date("2011-11-30"),54.0)

}
