% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET_Penman_Monteith.R
\name{PET.PM.FAO}
\alias{PET.PM.FAO}
\title{Penman-Monteith}
\usage{
PET.PM.FAO(
  JD,
  lat,
  Ele,
  tmin,
  tmax,
  wind,
  Relhum,
  Rs = NULL,
  wind.height = 2,
  Rs_Unit = "MJm2day"
)
}
\arguments{
\item{JD}{Julian day}

\item{lat}{Latitude in decimal degrees}

\item{Ele}{Elevation in meter}

\item{tmin}{Daily minimum temperature in degree Celsius}

\item{tmax}{Daily maximum temperature in degree Celsius}

\item{wind}{Wind speed in m/sec}

\item{Relhum}{Relative humidity in ratio}

\item{Rs}{Mean daily solar radiation. If Rs=NULL, Rs is esimated in MJ/m2/day.If Rs_unit is not 'MJm2day' but 'W/m2', Rs_Unit should be "Wm2'}

\item{wind.height}{Height of wind speed measurement above the ground surface in meter (default=2)}

\item{Rs_Unit}{Unit of Rs. If Rs_Unit='Wm2', it converts to MJ/m2/day. Default is Rs_Unit='MJm2day'.}
}
\description{
Calculation of the the Penman-Monteith potential evapotraspiration (PET) estimation (FAO-56 method)
}
\references{
Zotarelli, L., Dukes, M.D., Romero, C.C., Migliaccio, K.W., 2010. Step by Step Calculation of the Penman-Monteith Evapotranspiration (FAO-56 Method). IFAS Extension, University of Florida. .
Allen, R.G., Pereira, L.S., Raes, D., Smith, M., 1998. Crop evapotranspiration-Guidelines for computing crop water requirements-FAO Irrigation and drainage paper 56. Fao, Rome 300, D05109.
}
\author{
Hyung Eum (hyung.eum@gov.ab.ca)
}
