% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Priestley_Taylor.R
\name{PET_Priestley_Taylor}
\alias{PET_Priestley_Taylor}
\title{Priestley-Taylor PET method}
\usage{
PET_Priestley_Taylor(
  JD,
  lat,
  Ele,
  tmin,
  tmax,
  Relhum,
  Rs = NULL,
  Rs_Unit = "MJm2day"
)
}
\arguments{
\item{JD}{Julian day}

\item{lat}{Latitude in degree decimal (ex: 56.1)}

\item{Ele}{Elevation above mean sea level in meter}

\item{tmin}{Daily minimum temperature}

\item{tmax}{Daily maximum temperature}

\item{Relhum}{Relative humidity in ratio}

\item{Rs}{Mean daily solar radiation. If Rs=NULL, Rs is esimated in MJ/m2/day.If Rs_unit is not 'MJm2day' but 'W/m2', Rs_Unit should be "Wm2'}

\item{Rs_Unit}{Unit of Rs. If Rs_Unit='Wm2', it converts to MJ/m2/day. Default is Rs_Unit='MJm2day'.}
}
\value{
Daily PET in mm/day
}
\description{
Estimating daily PET based on using the Priestley-Taylor PET estimation method
}
