% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_PT.R
\name{PET_PT_AF}
\alias{PET_PT_AF}
\title{Simplified Priestley-Taylor PET method}
\usage{
PET_PT_AF(JD, lat, Elev = 700, tmin, tmax)
}
\arguments{
\item{JD}{Julian day}

\item{lat}{Latitude in degree decimal (ex: 56.1)}

\item{Elev}{Elevation above mean sea level in meter}

\item{tmin}{Daily minimum temperature}

\item{tmax}{Daily maximum temperature}
}
\value{
Daily PET in mm/day
}
\description{
Estimating daily PET based on min/max temperature using the simplified Priestley-Taylor PET estimation method
Solar radiation and net radiation are calculated by an empirical equation suggested by Alberta Climate Information Service.
}
\examples{
JD=Julian_Day(as.Date("2011-11-30"))
PET_PT_AF(JD=JD,lat=54.0,Elev=700,tmin=15,tmax=25)

}
\references{
Alberta Climate Information Service (ACIS). 2014. Estimating solar radiation using daily maximum and minimum temperature. Alberta Agriculture and Forestry.
Edmonton, AB. https://www.acis.alberta.ca/acis/docs/Estimating-solar-radiation-using-daily-max-and-min-temperatures-data-y2014_m06_d13.pdf
}
