% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentialSolar.R
\name{PotentialSolar}
\alias{PotentialSolar}
\title{Extraterrestrial solar radiation}
\usage{
PotentialSolar(lat, Jday)
}
\arguments{
\item{lat}{Latitude in radian}

\item{Jday}{Julian date or day of the year}
}
\value{
Potential solar radiation at the edge of the atmosphere \link{kJ m-2 d-1}
}
\description{
Calculating potential solar radiation given day of the year and latitude
}
\references{
Allen, R.G., Pereira, L.S., Raes, D., Smith, M., 1998. Crop evapotranspiration-Guidelines for computing crop water requirements-FAO Irrigation and drainage paper 56. Fao, Rome 300, D05109.
}
