% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slopefactor.R
\name{slopefactor}
\alias{slopefactor}
\title{Slopefactor}
\usage{
slopefactor(lat, Jday, slope, aspect)
}
\arguments{
\item{lat}{latitdue \link{rad}}

\item{Jday}{Julian date or day of the year \link{day}}

\item{slope}{slope of the ground \link{rad}}

\item{aspect}{ground aspect \link{rad from north}}
}
\value{
SF: Slope factors
}
\description{
Adjusts solar radiation for land slope and aspect relative to the sun, 1=level ground
}
\examples{
slopefactor(57.0,365,0.05,1)
}
