% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solar.R
\name{Solar}
\alias{Solar}
\title{Solar radiation estimation}
\usage{
Solar(
  lat,
  Jday,
  Tx,
  Tn,
  albedo = 0.2,
  forest = 0,
  slope = 0,
  aspect = 0,
  units = "MJm2d"
)
}
\arguments{
\item{lat}{Latitude in degrees}

\item{Jday}{Julian date}

\item{Tx}{Daily maximum temperature in degree C}

\item{Tn}{Daily minimum temperature in degree C}

\item{albedo}{Albedo (default=0.2)}

\item{forest}{Forest ratio}

\item{slope}{Slope derived from DEM}

\item{aspect}{Aspect derived from DEM}

\item{units}{Solar radiation unit (default=MJm2d)}

\item{latUnits}{Unit of latitude: 'degrees' or 'radians' (default=degrees)}

\item{printWarn}{logical (TRUE or FALSE)}
}
\value{
Solar radiation in units (MJ/m2/day or W/m2)
}
\description{
Estimating daily solar radiation in KJ/m2/day or W/m2
}
