% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solarangle.R
\name{solarangle}
\alias{solarangle}
\title{Solar angle}
\usage{
solarangle(lat, Jday)
}
\arguments{
\item{lat}{Latitdue \link{rad}}

\item{Jday}{Julian date or day of the year \link{day}}
}
\value{
Solar angle in radians
}
\description{
Angle of solar inclination from horizontal at solar noon \link{rad}
}
