% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaporPressure.R
\name{SVP.ClaCla}
\alias{SVP.ClaCla}
\title{Saturation vapor pressure by the Clausius-Clapeyron equation}
\usage{
SVP.ClaCla(Tair)
}
\arguments{
\item{Tair}{Air temperature in degree Celcius}
}
\value{
Saturation vapor pressure in hectopascal (hPa) or millibar (mb)

Saturation vapor pressure \link{hPa}
}
\description{
calculate saturation vapor pressure \eqn{E_s} at temperature \eqn{Tair}
}
\examples{
SVP.ClaCla(Tair=25.0)
}
\references{
Shaman, J., & Kohn, M., 2009, Absolute humidity modulates influenza survival, transmission, and seasonality. Proceedings of the National Academy of Sciences, 106(9), 3243-3248.
Wallace, J. M., & Hobbs, P. V., 2006, Atmospheric science: an introductory survey. 92, Academic press.
Specific gas constant of water vapor: https://en.wikipedia.org/wiki/Gas_constant#Specific_gas_constant
Latent heat of water vapor: https://en.wikipedia.org/wiki/Latent_heat
}
