% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transmissivity.R
\name{transmissivity}
\alias{transmissivity}
\title{Atmospheric Transmissivity}
\usage{
transmissivity(Tx, Tn, A = 0.75, C = 2.4, opt = "1day", JD = NULL)
}
\arguments{
\item{Tx}{maximum daily temperature \link{C}}

\item{Tn}{minimum daily temperature \link{C}}

\item{A}{Coefficient, 0.75}

\item{C}{Empirical coefficient, 2.4}

\item{JD}{}
}
\description{
Fraction of direct solar radiation passing through the atmosphere
}
\details{
Atmospheric Transmissivity is estimated by the Bristow-Campbell eqn, AT=A\link{1-exp(B(Tx-Tn)^C)}
B is an empirical coefficient and estimated by 0.036\emph{exp(-0.154}(Tx-Tn))
This function is used to estimate solar radiation (refer to 'Solar').
}
\references{
Bohn, T.J., Livneh, B., Oyler, J.W., Running, S.W., Nijssen, B., Lettenmaier, D.P., 2013. Global evaluation of MTCLIM and related algorithms for forcing of ecological and hydrological models. Agricultural and Forest Meteorology 176, 38–49. https://doi.org/10.1016/j.agrformet.2013.03.003
Bristow, K.L., Campbell, G.S., 1984. On the relationship between incoming solar radiation and daily maximum and minimum temperature. Agric. Forest Meteor. 31(2), 159–166.
Ndlovu, L.S., 1994. Weather Data Generation and Its Use in Estimating Evapotranspiration. PhD Thesis. Department of Biosystems Engineering (Engineering Science). Washington State University. Pullman, WA. 143pp.
}
