% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaporPressure.R
\name{WVP.RH.T}
\alias{WVP.RH.T}
\title{Partial water vapor pressure}
\usage{
WVP.RH.T(RH, Tair)
}
\arguments{
\item{RH}{relative humidity (ratio)}

\item{Tair}{temperature in degree celcius}
}
\value{
Partial water vapor pressure \link{hPa}
}
\description{
Calculate partial water vapor pressure given relative humdity and air temperature
}
